requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.1.0
requests>=2.31.0
boto3>=1.34.0
mcp>=1.0.0
bedrock-agentcore-starter-toolkit>=0.1.0
